requirements.txt
pymongo==4.8.0
pika==1.3.2
pydantic==2.8.2
PyYAML==6.0.1
yt-dlp==2024.7.25
ffmpeg-python==0.2.0
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
transformers==4.43.3
accelerate==0.33.0